refactor(web): extract shared FilesystemFields for storage forms
The partition and logical volume forms carried copy-paste identical
filesystem logic: volume template lookup, compatible filesystem
calculation, options building and the auto-reset effect. Move it to a
shared component driven by a single reusedDevice prop.

Each form keeps a thin wrapper that only resolves which existing device
is being reused, since that lives in form-specific fields (name vs
target) and collections (a drive's partitions vs a volume group's
logical volumes).

As a side effect, this solves a bug in the partition form: the
"Current" option description wrongly named the whole disk instead of
the partition being reused (e.g. "Do not format vdd and keep data"
instead of "Do not format vdd1 and keep data").
