+++ web/src/components/storage/logical-volume-form/FilesystemFields.tsx
@@ -20,129 +20,27 @@
  * find current contact information at www.suse.com.
  */
 
-import React, { useEffect, useMemo } from "react";
-import { unique, isEmpty } from "radashi";
+import React from "react";
 import { withForm } from "~/hooks/form";
-import FilesystemSelector from "~/components/storage/shared/FilesystemSelector";
-import { buildFilesystemOptions } from "~/components/storage/shared/helpers";
-import {
-  defaultOptions,
-  isReusingLogicalVolume,
-  FILESYSTEM_TYPE,
-  FILESYSTEM_ACTION,
-} from "./fields";
-import { useVolumeTemplate } from "~/hooks/model/system/storage";
-import { _ } from "~/i18n";
+import SharedFilesystemFields from "~/components/storage/shared/FilesystemFields";
+import { defaultOptions } from "./fields";
 
 import type { Storage as System } from "~/model/system";
-import type { ConfigModel } from "~/model/storage/config-model";
 
 type FilesystemFieldsProps = {
+  /**
+   * The volume group as it exists in the system, or undefined when it is new
+   * (a new volume group has no logical volumes to reuse).
+   */
   volumeGroup?: System.Device;
 };
 
-type FilesystemFieldsContentProps = {
-  volumeGroup?: System.Device;
-  target: string;
-  committedMountPoint: string;
-  filesystem: string;
-};
-
 /**
- * Inner component that renders the filesystem fields and handles auto-reset.
+ * Filesystem fields for the logical volume form.
  *
- * Mirrors the partition form behavior, resolving the reused device from the
- * volume group's logical volumes instead of a drive's partitions.
- *
- * ## Committed mount point pattern
- *
- * Uses `committedMountPoint` (not the live `mountPoint`) to derive filesystem
- * options and hints, avoiding misleading information while the user types an
- * incomplete value. See shared/MountPointField.tsx for the full pattern.
- *
- * ## Auto-reset behavior
- *
- * When the mount point changes to one that no longer supports the selected
- * filesystem type, the selector resets to "Default" so a compatible filesystem
- * can be chosen automatically. This is a UX convenience, not validation.
- */
-const FilesystemFieldsContent = withForm({
-  ...defaultOptions,
-  props: {
-    target: "",
-    committedMountPoint: "",
-    filesystem: "",
-  } as FilesystemFieldsContentProps,
-  render: function Render({ form, volumeGroup, target, committedMountPoint, filesystem }) {
-    const volume = useVolumeTemplate(committedMountPoint);
-    const defaultFilesystem = volume.fsType;
-    const isFallbackVolume = isEmpty(volume.mountPath);
-
-    const isReuse = isReusingLogicalVolume(target);
-    const selectedLogicalVolume = volumeGroup?.logicalVolumes?.find((lv) => lv.name === target);
-    const currentFsType = selectedLogicalVolume?.filesystem?.type;
-    const hasFilesystem = !!currentFsType;
-
-    // Memoized because it's used as a useEffect dependency below.
-    const usableFilesystems = useMemo(() => {
-      const volumeFilesystems = volume.outline.fsTypes || [];
-      return unique([defaultFilesystem, ...volumeFilesystems]);
-    }, [volume, defaultFilesystem]);
-
-    // The current filesystem can only be kept when it is compatible with the mount point.
-    const canKeepCurrentFilesystem =
-      hasFilesystem && usableFilesystems.includes(currentFsType as ConfigModel.FilesystemType);
-
-    const filesystemOptions = useMemo(
-      () =>
-        buildFilesystemOptions(usableFilesystems, {
-          device: selectedLogicalVolume,
-          canKeepCurrentFilesystem,
-        }),
-      [usableFilesystems, canKeepCurrentFilesystem, selectedLogicalVolume],
-    );
-
-    // Auto-reset filesystem to Default when it becomes incompatible with the mount point.
-    useEffect(() => {
-      if (filesystem === FILESYSTEM_TYPE.AUTO) return;
-      if (filesystem === FILESYSTEM_ACTION.REUSE && canKeepCurrentFilesystem) return;
-      if (usableFilesystems.includes(filesystem as ConfigModel.FilesystemType)) return;
-
-      form.setFieldValue("filesystem", FILESYSTEM_TYPE.AUTO);
-    }, [usableFilesystems, filesystem, form, canKeepCurrentFilesystem]);
-
-    return (
-      <>
-        <FilesystemSelector
-          form={form}
-          defaultFilesystem={defaultFilesystem}
-          committedMountPoint={committedMountPoint}
-          filesystemOptions={filesystemOptions}
-          usableFilesystems={usableFilesystems}
-          selectedDevice={isReuse && hasFilesystem ? selectedLogicalVolume : undefined}
-          isFallback={isFallbackVolume}
-        />
-
-        <form.AppField name="showMoreFilesystemSettings">
-          {(field) => (
-            <field.CheckboxField
-              label={
-                // TRANSLATORS: checkbox label for additional filesystem configuration options
-                _("Define more file system settings")
-              }
-            />
-          )}
-        </form.AppField>
-      </>
-    );
-  },
-});
-
-/**
- * Filesystem selection and label fields for the logical volume form.
- *
- * Thin wrapper that subscribes to the fields the inner content needs and
- * delegates to the shared FilesystemSelector.
+ * Thin wrapper around the shared FilesystemFields: it only resolves which
+ * existing logical volume is being reused, from the `target` field (empty
+ * string means a new logical volume) and the volume group's logical volumes.
  */
 const FilesystemFields = withForm({
   ...defaultOptions,
@@ -151,20 +49,11 @@ const FilesystemFields = withForm({
   } as FilesystemFieldsProps,
   render: function Render({ form, volumeGroup }) {
     return (
-      <form.Subscribe
-        selector={(s) => ({
-          target: s.values.target,
-          committedMountPoint: s.values.committedMountPoint,
-          filesystem: s.values.filesystem,
-        })}
-      >
-        {({ target, committedMountPoint, filesystem }) => (
-          <FilesystemFieldsContent
+      <form.Subscribe selector={(s) => s.values.target}>
+        {(target) => (
+          <SharedFilesystemFields
             form={form}
-            volumeGroup={volumeGroup}
-            target={target}
-            committedMountPoint={committedMountPoint}
-            filesystem={filesystem}
+            reusedDevice={volumeGroup?.logicalVolumes?.find((lv) => lv.name === target)}
           />
         )}
       </form.Subscribe>
